Very large performance update. The generic pool structure has been removed in favor of simple ArrayLists in most cases: the only two pools that require more attention are the constant and attribute pools. The constant pool has been readded to its own file, ConstantPool.java.
Under the assumption that the latter is read and appendable only: i.e. no
constants are ever removed, a large set of optimizations have been
employed. The ConstantPool struct holds an array of constant objects. The
constant object has been redefined to simply be a pointer to a specific
point in the constant pool, and hold an arbitrary value in the field
'complex'.

When reading the constant pool, we now fork the input stream to a writable
form after parsing and cache it: this means the constant pool need not be
recomputed from scratch. All new* constant generators write to the stream
when generating constants.

Heavy constants, or constants that require other constants to retrieve
their value have been defined as 'lazy': their index's 'resolved' flag is
set to false. These lazy constants include all but the primitive constant
types. This optimization results in that operations with a large overhead,
such as UTF-8 string decoding can be saved up until they are required.
Under the stipulation that most such constants will never be accessed,
this optimization causes dozens of milliseconds of speed improvement.
While this does mean an initial latency upon accessing a constant, the
computed value is cached in the 'complex' field of the index.

Dump/read methods have also been removed from most structures, with the
exceptions of attributes. This means a cleaner interface for users to use.
Additionally, since member parsing etc. is done in the main body of the
class, new methods can be defined and JBL will compute the constant pool
entries for names, descriptors, strings and such. This results in an
extreme ease of adding new members or defining new implementing
interfaces: no (or very minimal) constant pool modification has to be done
by hand.

The Bytes class has also been removed, in favor of an updated ByteStream.

The new ByteStream is purely based off of operations on a byte buffer. The
dependency on DataInputStream and DataOutputStream have been consequently
removed. Hence, the calls to ByteStream.read and .write no longer have to
be wrapped in try/catch, once again speeding execution time up.

the 'core' and 'visitor' package have been moved to zip files for the
simple reason that the subtle changes in API have rendered them broken,
and fixing them is out of the current scope. They will be fixed when this
branch is merged back to master.